S-44: fixed relative paths after refactor
